chore: delete Alpine editor remnants and Phase 0 spike
Phase 1.14 — completes the React migration's Phase 1 goal. After this
lands, the visual editor tree contains no Alpine-era or spike-era
scaffolding.

- Remove resources/js/editor-spike/ (Phase 0 architecture spike)
- Remove resources/js/visual-editor/editor/__dnd-kit-spike__/ (Phase 1
  dnd-kit harness that depended on @spike/richtext/useTiptap)
- Remove editor-spike route, blade view, index.html entry, and the
  visual-editor-spike-assets publish tag
- Consolidate vite.editor.config.ts into the root vite.config.ts and
  drop the spike-only vite.config.ts; simplify npm scripts
- Drop the @spike path alias from tsconfig.json and vitest.config.ts
- Move the Tiptap/jsdom test setup into
  resources/js/visual-editor/editor/test-setup.ts
- Update EditorMountViewTest route assertion and remove
  public/editor-spike/ from .gitignore
- Mark Phase 1 as landed in REACT-MIGRATION.md

Closes #276
