--- index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Visual Editor Spike</title>
-    </head>
-    <body>
-        <div id="editor-spike-root"></div>
-        <script type="module" src="/resources/js/editor-spike/main.tsx"></script>
-    </body>
-</html>
