--- resources/js/editor-spike/blocks/paragraph/index.ts
@@ -1,13 +1,0 @@
-import { registerBlock } from '../../registry';
-import ParagraphEdit from './edit';
-
-export const PARAGRAPH_BLOCK_NAME = 've/paragraph';
-
-export function registerParagraphBlock(): void {
-    registerBlock({
-        name: PARAGRAPH_BLOCK_NAME,
-        edit: ParagraphEdit,
-    });
-}
-
-export { default as ParagraphEdit } from './edit';
