--- resources/js/visual-editor/editor/__dnd-kit-spike__/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>dnd-kit spike — visual-editor (dev only)</title>
-    </head>
-    <body>
-        <div id="spike-root"></div>
-        <script type="module" src="./main.tsx"></script>
-    </body>
-</html>
