--- resources/js/visual-editor/editor/__dnd-kit-spike__/spike.css
@@ -1,83 +1,0 @@
-:root {
-    color-scheme: light;
-    font-family: system-ui, -apple-system, sans-serif;
-    color: #111;
-    background: #f5f5f5;
-}
-
-body {
-    margin: 0;
-    padding: 2rem;
-    background: #f5f5f5;
-    color: #111;
-}
-
-.spike-root {
-    max-width: 720px;
-    margin: 0 auto;
-}
-
-.spike-header h1 {
-    margin: 0 0 0.5rem;
-    font-size: 1.25rem;
-}
-
-.spike-header p {
-    margin: 0 0 1.5rem;
-    color: #555;
-    font-size: 0.9rem;
-}
-
-.spike-list {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 0.75rem;
-}
-
-.spike-block {
-    display: flex;
-    align-items: stretch;
-    gap: 0.5rem;
-    background: #fff;
-    border: 1px solid #ddd;
-    border-radius: 6px;
-    padding: 0.5rem;
-}
-
-.spike-block__handle {
-    flex: 0 0 auto;
-    width: 28px;
-    border: none;
-    background: #eee;
-    border-radius: 4px;
-    cursor: grab;
-    font-size: 1.1rem;
-    color: #666;
-    touch-action: none;
-}
-
-.spike-block__handle:active {
-    cursor: grabbing;
-}
-
-.spike-block__handle:focus-visible {
-    outline: 2px solid #3b82f6;
-    outline-offset: 2px;
-}
-
-.spike-block__body {
-    flex: 1 1 auto;
-    padding: 0.25rem 0.5rem;
-}
-
-.ve-richtext {
-    outline: none;
-    min-height: 1.5rem;
-}
-
-.ve-richtext p {
-    margin: 0;
-}
