--- resources/views/editor-spike.blade.php
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Visual Editor Spike</title>
-    <link rel="stylesheet" href="{{ asset('vendor/visual-editor/editor-spike/main.css') }}">
-</head>
-<body>
-    <div id="editor-spike-root"></div>
-    <script type="module" src="{{ asset('vendor/visual-editor/editor-spike/main.js') }}"></script>
-</body>
-</html>
