+++ routes/web.php
@@ -4,10 +4,6 @@ declare(strict_types=1);
 
 use Illuminate\Support\Facades\Route;
 
-Route::get('/editor-spike', function () {
-    return view('visual-editor::editor-spike');
-})->name('visual-editor.editor-spike');
-
 Route::get('/editor', function () {
     return view('visual-editor::editor.mount', [
         'postId' => '1',
