+++ tsconfig.json
@@ -19,12 +19,10 @@
         "types": ["vitest/globals", "@testing-library/jest-dom"],
         "baseUrl": ".",
         "paths": {
-            "@spike/*": ["resources/js/editor-spike/*"],
             "@editor/*": ["resources/js/visual-editor/editor/*"]
         }
     },
     "include": [
-        "resources/js/editor-spike/**/*",
         "resources/js/visual-editor/**/*"
     ],
     "exclude": ["node_modules", "vendor", "public", "dist"]
