+++ vitest.config.ts
@@ -6,15 +6,14 @@ export default defineConfig({
     plugins: [react()],
     resolve: {
         alias: {
-            '@spike': resolve(__dirname, 'resources/js/editor-spike'),
+            '@editor': resolve(__dirname, 'resources/js/visual-editor/editor'),
         },
     },
     test: {
         globals: true,
         environment: 'jsdom',
-        setupFiles: ['./resources/js/editor-spike/test-setup.ts'],
+        setupFiles: ['./resources/js/visual-editor/editor/test-setup.ts'],
         include: [
-            'resources/js/editor-spike/**/*.{test,spec}.{ts,tsx}',
             'resources/js/visual-editor/**/*.{test,spec}.{ts,tsx}',
         ],
     },
